fix(issue-72): phase 3 priority 3 - shorten UI and hook long functions to under 150 lines (#85)
* fix(issue-72): phase 3 priority 3 - shorten UI and hook long functions to under 150 lines

- useAIChat: extract executeSendMessage to useAIChatExecute.ts
- SearchResults: extract SearchResultCard, SearchResultsLoadingSkeleton, SearchResultsEmptyState
- HeaderSearchBar: extract handleSearchKeyDown and HeaderSearchDropdownContent
- AIChatInput: split into aiChatInputHelpers, useAIChatInput (useAIChatInputChips)
- useImageUploadManager: extract runSingleUpload, editor node helpers, filterImageFiles; unify processImageUpload

Ref: #72

* fix: enhance AI chat input and editor lifecycle handling

- Added a check to ensure the editor contains the selection range in `replaceMentionWithChip`.
- Refactored `useAIChatInput` to utilize `pendingRefs` and `isEmpty` directly, improving state management.
- Updated `useEditorLifecycle` to correctly identify empty documents based on ProseMirror's node size.
- Improved error handling in `executeSendMessage` and `runSingleUpload` for better user feedback.
- Adjusted various components for improved accessibility and UI consistency.


---------
